pyproject.toml
[project]
name = "user-management"
version = "0.1.0"
description = "Multi-tenant user management backend"
authors = [
    {name = "PlanckScale Team", email = "dev@planckscale.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "motor>=3.3.0",
    "redis>=5.0.0",
    "pydantic[email]>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.24.1",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "requests>=2.32.3",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
user-management = "users.main:main"